feat(share): feed publish
add feed transcription selection view
